requirements.txt
langchain==0.1.12
langchain-openai==0.0.8
python-dotenv==1.0.1
requests==2.31.0
beautifulsoup4==4.12.3
googlesearch-python==1.2.3
langchainhub